pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neoscene"
version = "0.1.0"
description = "Neoscene is a text-to-scene generator for MuJoCo."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Neoscene Team"}
]

dependencies = [
    "mujoco",
    "pydantic",
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "pyyaml",
    "google-generativeai",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
]

[project.scripts]
neoscene = "neoscene.app.main:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

